removes morecore(), moves everything to size_t, adds stats.c

+++ malloc_j/src/calloc_j.c
@@ -1,8 +1,8 @@
 #include <string.h>
 #include "malloc_j.h"
 
-void* calloc_j(const unsigned int size, const unsigned int count) {
-  unsigned int total_bytes = size * count;
+void* calloc_j(const size_t size, const size_t count) {
+  size_t total_bytes = size * count;
   void* space = malloc_j(total_bytes);
   memset(space, '\0', total_bytes);
   return space;
